Adaugare Puncte de lucru

+++ frontend/controllers/MinistereExceptiiController.php
@@ -31,33 +31,30 @@ class MinistereExceptiiController extends Controller
         );
     }
 
-    /**
-     * Lists all MinistereExceptii models.
-     *
-     * @return string
-     */
     public function actionIndex()
     {
-        $searchModel = new MinistereExceptiiSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin_minister')){
+            $searchModel = new MinistereExceptiiSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+        return $this->redirect(['site/index']);
+
+
     }
 
-    /**
-     * Displays a single MinistereExceptii model.
-     * @param int $id ID
-     * @return string
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin_minister')){
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
+        return $this->redirect(['site/index']);
     }
 
     /**
@@ -67,19 +64,22 @@ class MinistereExceptiiController extends Controller
      */
     public function actionCreate()
     {
-        $model = new MinistereExceptii();
+        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin_minister')){
+            $model = new MinistereExceptii();
 
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $model->save()) {
+                    return $this->redirect(['view', 'id' => $model->id]);
+                }
+            } else {
+                $model->loadDefaultValues();
             }
-        } else {
-            $model->loadDefaultValues();
-        }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+        }
+        return $this->redirect(['site/index']);
     }
 
     /**
@@ -91,15 +91,19 @@ class MinistereExceptiiController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin_minister')){
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
+        return $this->redirect(['site/index']);
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -111,9 +115,12 @@ class MinistereExceptiiController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin_minister')){
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
+        return $this->redirect(['site/index']);
     }
 
     /**
